Refactor for vertical adjustments
